+++ template/ExtensionIndex.php
@@ -47,14 +47,14 @@
             </ul>
         </td>
         <td>
-            <?php if ($extension['manifest']['version']) : ?>
+            <?php if (! empty($extension['manifest']['version'])) : ?>
                 <a href='<?php $this->url("Extension/changelog?id_extension=$id_e") ?>'><?php hecho($extension['manifest']['version']); ?></a>
             <?php else :?>
                 <span class='text_alert'>NON VERSIONNÉE</span>
             <?php endif;?>
             &nbsp;
-            (<?php hecho($extension['manifest']['revision'])?>)
-            <?php if ($extension['manifest']['autre-version-compatible']) : ?>
+            (<?php hecho($extension['manifest']['revision'] ?? "")?>)
+            <?php if (! empty($extension['manifest']['autre-version-compatible'])) : ?>
                 <br/>Versions compatibles :
                 <ul>
                     <?php foreach ($extension['manifest']['autre-version-compatible'] as $version) : ?>
@@ -65,7 +65,7 @@
 
         </td>
         <td>
-            <?php hecho($extension['manifest']['pastell-version'])?>
+            <?php hecho($extension['manifest']['pastell-version'] ?? '')?>
             <?php if (! $extension['pastell-version-ok']) : ?>
             &nbsp;-&nbsp;<span class='text_alert'>KO</span>
             <?php endif;?>